… client/go/go.sum