Gives a decent boost over stdlib:

```
$ benchstat /tmp/old.txt /tmp/new.txt
goos: darwin
goarch: arm64
pkg: github.com/vespa-engine/vespa/client/go/internal/vespa/document
                        │ /tmp/old.txt │            /tmp/new.txt             │
                        │    sec/op    │   sec/op     vs base                │
ClientSendCompressed-10   13.243µ ± 1%   4.701µ ± 1%  -64.50% (p=0.000 n=10)
```

This library also provides a pure-Go implementation of zstd for when we want to
add that.

Review, but hold merge.

@jonmv