@geirst 2023-05-05
For better visibility in such cases, consider using the label "DO NOT MERGE".